requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
line-bot-sdk==1.16.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
soupsieve==2.0

